+++ crates/rslint_cli/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added 
+
+- Added ways of configuring the formatter used through CLI and config
+- Made all fields of config public
+
 ## [0.2.1] - 2020-10-21
 
 ### Changed